(jbutkus) How Come The Bank is 404
This talk is likely to span 30-45 minutes, while covering the following topics:
  * what happens when one uses a payment card at a POS/ATM;
  * an abridged version of responsibilities of all parties involved in making the payment succeed or fail;
  * how can the owner of a payment card (no, very likely, you are not the owner -- you are only the holder
    of the card, while some bank-link institutioin is the owner) vanish, move elsewhere, and what happens
    to the card and pending payments at that time.

Save for some portions which better remain private for the sake of the companies which are still very active,
the talk is based on an actual experience doing a `308 Permanent Redirect` of sorts. Over a span of 6 hours.
And then again, over a similar time period, that time succeeding more.