Apnea diving 101
Did you ever want to dive to an ocean, sea or so and see what's under,
but without all the bulky gear that keeps us alive by providing oxygen?
Just get goggles and here we go! Sounds simple!

There is such a cruel sport, or leisure activity, depending on how you
see the world, called "apnea diving". That is, to plain Lithuanian,
"diving without supplemental oxygen". In this talk I will go through:

- some apnea diving trivia.
- how our body reacts to (lack of) oxygen.
- how much can we push it before it's not safe?
- how to get started?

I will bring equipment and do some on-the-spot experiments with a
volunteer or two.