Add @ViliusKraujutis presentation for 2023
Hotwiring Cold Science: A Journey into a Freeze Drier's Brain is an exploration of reverse engineering and troubleshooting a faulty freeze drier. In this presentation, Vilius Kraujutis shares his intriguing adventure of dissecting a "Lyo-Tech CHWC-20A" freeze dryer that stopped working after its warranty period. He takes the audience on a journey through the inner workings of the device, uncovering the role of each component such as the chamber, warming shelves, sensors, and control system, and their interplay in the fascinating process of lyophilization. Despite a lack of prior expertise in analog electronics, Vilius managed to navigate the complex circuits and systems, resulting in a humorous and enlightening narrative about understanding the freeze drying process, diagnosing issues, and even finding a fix for the faulty heating system. The presentation is not only about the technical aspects but also an exciting exploration of scientific concepts like sublimation, deposition, and the triple point of water. This is an engaging, thought-provoking session that brings science, engineering, and curiosity together in an accessible and entertaining manner.