Update 404.txt
Although we are not often able to put the sense in words, in a work-related environment, the interpersonal dynamics work differently depending on with whom and why we join the efforts. In this presentation, Il'd like to look closer at what are the differences between teams and groups and challenges related to decision making in such environments. Inspired by workplace transition from small company to small enterprise.